+++ internal/random_bytes.mjs
@@ -12,5 +12,5 @@ export default async function random_bytes(bytes = 32) {
   if (typeof window == "undefined") {
     const { randomBytes } = await import("crypto");
     return Uint8Array.from([...randomBytes(bytes)]);
-  } else return crypto.getRandomValues(new Uint8Array(bytes));
+  } else return window.crypto.getRandomValues(new Uint8Array(bytes));
 }