+++ sign_packed_txn.mjs
@@ -1,4 +1,4 @@
-// @ts-check
+import sha256 from "universal-sha256-js/sha256.mjs";
 
 import sign_tnx from "./sign_txn.mjs";
 
@@ -11,30 +11,24 @@ import sign_tnx from "./sign_txn.mjs";
  * @param {String} PackedTxn.transaction_header Serialised transaction header.
  * @param {String} PackedTxn.transaction_body Serialised transaction body.
  * @param {String} PackedTxn.wif_private_key Private key (wallet import format).
+ * @param {String} [PackedTxn.extension] Extension
  * @returns {Promise<String>} Signature
- * @example <caption>Usage `sign_packed_txn`.</caption>
- * ```js
- * import sign_packed_txn from 'antelope-ecc/sign_packed_txn.mjs'
- *
- * sign_packed_txn(
- *  {
- *    chain_id: '2a02a0053…',
- *    transaction_header: 'fa123232…',
- *    transaction_body: 'fa45ffa2…',
- *    wif_private_key: '5f…'
- *  }
- *)
- * ```
- * The logged output was SIG_K1_….
  */
 export default async function sign_packed_txn({
   chain_id,
   transaction_header,
   transaction_body,
   wif_private_key,
+  extension = "0000000000000000000000000000000000000000000000000000000000000000",
 }) {
   return sign_tnx({
-    hex: chain_id + transaction_header + transaction_body,
+    hash: await sha256(
+      Uint8Array.from(
+        String(chain_id + transaction_header + transaction_body + extension)
+          .match(/[a-fA-F0-9]{2}/gmu)
+          .map((i) => Number(`0x${i}`))
+      )
+    ),
     wif_private_key,
   });
 }