+++ test/legacy_to_public.test.mjs
@@ -1,9 +1,9 @@
-import { deepStrictEqual } from "assert";
+import { deepStrictEqual, throws } from "assert";
 
-import legacy_to_private_key from "../legacy_to_private_key.mjs";
-import legacy_to_public_key from "../legacy_to_public_key.mjs";
-import private_key_to_legacy from "../private_key_to_legacy.mjs";
-import public_key_to_legacy from "../public_key_to_legacy.mjs";
+import legacy_to_private_key from "../keys/legacy_to_private_key.mjs";
+import legacy_to_public_key from "../keys/legacy_to_public_key.mjs";
+import private_key_from_wif from "../keys/private_key_from_wif.mjs";
+import public_key_from_wif from "../keys/public_key_from_wif.mjs";
 
 export default (tests) => {
   tests.add("legacy key to public key", async () => {
@@ -11,38 +11,38 @@ export default (tests) => {
       await legacy_to_public_key(
         "EOS53jowyaGC1WrYJefSHTTmGvZcySUFkEpmCDmEd8txunDChput7"
       ),
-      "PUB_K1_53jowyaGC1WrYJefSHTTmGvZcySUFkEpmCDmEd8txunDCqCCVR",
+      await public_key_from_wif(
+        "PUB_K1_53jowyaGC1WrYJefSHTTmGvZcySUFkEpmCDmEd8txunDCqCCVR"
+      ),
       "Expected converted key to be PUB_K1_53jowyaGC1WrYJefSHTTmGvZcySUFkEpmCDmEd8txunDCqCCVR"
     );
-
-    deepStrictEqual(
-      "EOS53jowyaGC1WrYJefSHTTmGvZcySUFkEpmCDmEd8txunDChput7",
-      await public_key_to_legacy(
-        "PUB_K1_53jowyaGC1WrYJefSHTTmGvZcySUFkEpmCDmEd8txunDCqCCVR"
-      )
-    );
-
-    deepStrictEqual(
-      await legacy_to_private_key(
-        "5KML6yCUABWYxuEexgMZPJA9641SptvHdB5Gm5KZW8rFeGf5uak"
-      ),
-      "PVT_K1_2Y3XHkP5iwZhtrNvUufJFR1sTBXcm4CuN1VXuGpGFzcUa8vu23",
-      "Expected converted private key."
-    );
-    deepStrictEqual(
-      await private_key_to_legacy(
-        "PVT_K1_2Y3XHkP5iwZhtrNvUufJFR1sTBXcm4CuN1VXuGpGFzcUa8vu23"
-      ),
-      "5KML6yCUABWYxuEexgMZPJA9641SptvHdB5Gm5KZW8rFeGf5uak",
-      "Expected converted private key."
-    );
-
     deepStrictEqual(
       await legacy_to_public_key(
         "EOS6p9CLAjviu2kMTCRNSh85EmutrH5X2PeBwt4YHKQ1fiJLZDEtZ"
       ),
-      "PUB_K1_6p9CLAjviu2kMTCRNSh85EmutrH5X2PeBwt4YHKQ1fiJMHvUqF",
+      public_key_from_wif(
+        "PUB_K1_6p9CLAjviu2kMTCRNSh85EmutrH5X2PeBwt4YHKQ1fiJMHvUqF"
+      ),
       "Expected converted key to be PUB_K1_6p9CLAjviu2kMTCRNSh85EmutrH5X2PeBwt4YHKQ1fiJMHvUqF"
     );
+    deepStrictEqual(
+      await legacy_to_private_key(
+        "5KML6yCUABWYxuEexgMZPJA9641SptvHdB5Gm5KZW8rFeGf5uak"
+      ),
+      await private_key_from_wif(
+        "PVT_K1_2Y3XHkP5iwZhtrNvUufJFR1sTBXcm4CuN1VXuGpGFzcUa8vu23"
+      )
+    );
+    //   "Expected converted private key."
+    deepStrictEqual(
+      private_key_from_wif(
+        "PVT_K1_2Y3XHkP5iwZhtrNvUufJFR1sTBXcm4CuN1VXuGpGFzcUa8vu23"
+      ),
+      legacy_to_private_key(
+        "5KML6yCUABWYxuEexgMZPJA9641SptvHdB5Gm5KZW8rFeGf5uak"
+      ),
+      "Expected converted private key."
+    );
+    throws(() => public_key_from_wif("NOT_PUBK1_KEY"));
   });
 };