+++ test/random_bytes.test.mjs
@@ -1,7 +1,7 @@
 import assert from "assert";
 import crypto from "crypto";
 
-import random_bytes from "../random_bytes.mjs";
+import random_bytes from "../internal/random_bytes.mjs";
 
 const { ok } = assert;
 
@@ -10,8 +10,7 @@ export default (tests) => {
     const bytes_web = await random_bytes();
     ok(bytes_web.length == 32);
 
-    global.window = true;
-    global.crypto = crypto.webcrypto;
+    global.window = { crypto: crypto.webcrypto };
 
     const bytes = await random_bytes();
     ok(bytes.length == 32);